reinforcement/requirements.txt
# Core ML/Data Libraries
torch
numpy
pandas
faiss-cpu

# NLP Libraries
nltk
gensim

# Storage/Database
minio
psycopg2-binary
python-dotenv

# Settings Management
pydantic
pydantic-settings

# Logging/Utils
python-json-logger